# public/postman/schemas/api.yml
openapi: 3.0.3
info:
  title: Arcane City
  version: 2025.06.09
  description: REST API for working with Arcane City data
  contact:
    name: Geoff Maddock
    email: geoff.maddock@gmail.com
servers:
  - url: https://arcane.city
    description: Production
  - url: https://stage.arcane.city
    description: Stage
  - url: https://dev.arcane.city
    description: Development
...
components:
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          maxLength: 255
          description: A name of a thread
          example: Forum
        slug:
          type: string
          maxLength: 255
          description: A unique identifier name for the post in kebab-case
          example: post-slug
        description:
          type: string
          example: An extensive breakdown of the post
          description: An extensive breakdown of the post
        visibility_id:
          type: integer
          readOnly: true
          example: 1
          description: Relation to the visibility table that defines the visibility of the post
        thread_id:
          type: integer
          readOnly: true
          example: 1
          description: Relation to the thread table that defines the thread of the post
        created_at:
          type: string
          example: "2018-03-20T09:12:28Z"
          format: date-time
          description: Date and time that the entity type was created
          readOnly: true
    Posts:
      allOf: [$ref: "#/components/schemas/Pagination"]
      type: object
      properties:
        data:
          type: array
          description: List of the current page of posts
          items:
            "$ref": "#/components/schemas/Post"
...
    Pagination:
      type: object
      properties:
        current_page:
          description: The current page in the paginated list
          type: integer
        first_page_url:
          description: URL for the first page
          type: string
          example: "https://dev.arcane.city/api/resources?page=1"
        from:
          description: First element in the current page
          type: integer
          example: 1
        last_page:
          description: Last page of results
          type: integer
          example: 1
        last_page_url:
          description: URL for the first page
          type: string
          example: "https://dev.arcane.city/api/resources?page=100"
        links:
          description: Array of links used to paginate the list of entities
          type: array
          items: {}
        next_page_url:
          description: URL for the next page
          type: string
          example: "https://dev.arcane.city/api/resources?page=2"
        path:
          description: Path for the primary route
          type: string
          example: "https://dev.arcane.city/api/resources"
        per_page:
          description: Number of results per page
          type: integer
          example: 25
        prev_page_url:
          description: URL for the previous page
          type: string
          example: "https://dev.arcane.city/api/resources?page=0"
        to:
          description: Last element in the current page
          type: integer
          example: 25
        total:
          description: Total number of results from the request
          type: integer
          example: 100
        data:
          description: List of the current page of the paginated entity
          type: array
          items: {}
security:
  - basicAuth: []
paths:
...
  /api/posts:
    get:
      tags:
        - posts
      summary: Get Posts
      operationId: getPosts
      parameters:
        - name: filters[name]
          in: query
          required: false
          description: A filter query of the post name
          schema:
            type: string
            example: Post name
        - name: limit
          in: query
          required: false
          description: The limit on the number of results to return per page
          schema:
            type: integer
            example: 25
        - name: page
          in: query
          required: false
          description: The page number to return
          schema:
            type: integer
            example: 1
        - name: sort
          in: query
          required: false
          description: A column to be used to sort the query
          schema:
            type: string
            example: name
        - name: direction
          in: query
          required: false
          description: A string indicating the sort direction of the query (asc or desc)
          schema:
            type: string
            example: asc
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Posts"
...
  /api/threads/{threadId}/posts:
    get:
      parameters:
        - name: threadId
          description: The unique identifier of the thread
          in: path
          required: true
          schema:
            description: The unique identifier of a thread
            type: integer
            readOnly: true
            example: 1
      tags:
        - threads
      summary: Get Posts in Thread
      operationId: getPostsInThreadById
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Posts"
  /api/threads/{threadId}/posts/{postId}:
    parameters:
      - name: threadId
        description: The unique identifier of the thread
        in: path
        required: true
        schema:
          type: integer
          readOnly: true
          example: 1
      - name: postId
        description: The unique identifier of the post in the thread
        in: path
        required: true
        schema:
          type: integer
          readOnly: true
          example: 1
    get:
      parameters:
        - name: threadId
          description: The unique identifier of the thread
          in: path
          required: true
          schema:
            description: The unique identifier of a thread
            type: integer
            readOnly: true
            example: 1
      tags:
        - threads
      summary: Get Post in Thread
      operationId: getPostInThreadById
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
...